LPS-153858 Redirect Interval for 404 not allowing floats to set 30 minutes

+++ modules/apps/redirect/redirect-service/src/main/java/com/liferay/redirect/internal/configuration/RedirectConfiguration.java
@@ -53,6 +53,6 @@ public interface RedirectConfiguration {
 		description = "check-redirect-not-found-entries-interval-help",
 		name = "check-redirect-not-found-entries-interval", required = false
 	)
-	public int checkRedirectNotFoundEntriesInterval();
+	public float checkRedirectNotFoundEntriesInterval();
 
 }
